+++ benchmarks/Multi-Doc/InsertManySmallBenchmark.cs
@@ -1,13 +1,13 @@
 using System;
-using System.Collections.Generic;
-using BenchmarkDotNet.Attributes;
 using MongoDB.Bson;
 using MongoDB.Driver;
+using System.Collections.Generic;
+using BenchmarkDotNet.Attributes;
 using static benchmarks.BenchmarkExtensions;
 
 namespace benchmarks.Multi_Doc;
 
-
+[IterationCount(100)]
 [BenchmarkCategory("MultiBench", "WriteBench", "DriverBench")]
 public class InsertManySmallBenchmark
 {
@@ -19,7 +19,7 @@ public class InsertManySmallBenchmark
     [GlobalSetup]
     public void Setup()
     {
-        string mongoUri = Environment.GetEnvironmentVariable("MONGO_URI");
+        string mongoUri = Environment.GetEnvironmentVariable("MONGODB_URI");
         _client = mongoUri != null ? new MongoClient(mongoUri) : new MongoClient();
         _client.DropDatabase("perftest");
         var smallDocument = ReadExtendedJson("../../../../../../../data/single_and_multi_document/small_doc.json");
