+++ benchmarks/ParallelBench/MultiFileImportBenchmark.cs
@@ -1,35 +1,30 @@
 using System;
-using System.Collections.Generic;
 using System.IO;
 using System.Linq;
-using System.Threading.Tasks;
-using BenchmarkDotNet.Attributes;
 using MongoDB.Bson;
 using MongoDB.Driver;
+using System.Threading.Tasks;
+using BenchmarkDotNet.Attributes;
+using System.Collections.Generic;
 
 namespace benchmarks.ParallelBench;
 
-[IterationCount(2)]
-[WarmupCount(1)]
+[WarmupCount(2)]
+[IterationCount(4)]
 [BenchmarkCategory("ParallelBench", "WriteBench", "DriverBench")]
 public class MultiFileImportBenchmark
 {
     private MongoClient _client;
     private IMongoDatabase _database;
-    // private Action[] _parallelInserts;
     private IMongoCollection<BsonDocument> _collection;
-    // private BenchmarkThreadHelper _threadHelper;
 
     [GlobalSetup]
     public void Setup()
     {
-        string mongoUri = Environment.GetEnvironmentVariable("MONGO_URI");
+        string mongoUri = Environment.GetEnvironmentVariable("MONGODB_URI");
         _client = mongoUri != null ? new MongoClient(mongoUri) : new MongoClient();
         _client.DropDatabase("perftest");
         _database = _client.GetDatabase("perftest");
-        // _threadHelper = new BenchmarkThreadHelper();
-        // _parallelInserts = new Action[100];
-        // CreateTasks();
     }
 
     [IterationSetup]
@@ -48,7 +43,6 @@ public class MultiFileImportBenchmark
             tasks[i] = Task.Factory.StartNew(ImportFile(i));
         }
         Task.WaitAll(tasks);
-        // _threadHelper.ExecuteAndWait();
     }
 
     [GlobalCleanup]
@@ -56,15 +50,6 @@ public class MultiFileImportBenchmark
     {
         _client.DropDatabase("perftest");
     }
-
-    // private void CreateTasks()
-    // {
-    //     for (int i = 0; i < 100; i++)
-    //     {
-    //         // _parallelInserts[i] = ImportFile(i);
-    //         _threadHelper.Add(ImportFile(i));
-    //     }
-    // }
 
     private Action ImportFile(int fileNumber)
     {
